requirements.txt
Django==1.11.3
django-filter==1.0.4
djangorestframework==3.6.3
django-celery
celery==4.1.0
django-celery-results
gunicorn==19.6.0
psycopg2==2.6.2
twilio==5.6.0
slackclient==1.0.6
django-anymail==0.10
django-storages==1.6.3
boto==2.48.0
kong_oauth==0.1
django-phonenumber-field
python-dateutil
weasyprint

ipython
ipdb
responses

coverage
codacy-coverage
Faker
